requirements.txt
opencv-python
imread-from-url
onnxruntime
cap-from-youtube
pymongo
python-dotenv
boto3
web3
schedule
cloudinary
moviepy
vidgear